[core] Privatize math.hpp and vec.hpp
